% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_scale.r
\name{check_scale}
\alias{check_scale}
\title{check that this scale is present in Natural Earth}
\usage{
check_scale(x)
}
\arguments{
\item{x}{scale of map to return, one of \code{110}, \code{50}, \code{10}}
}
\value{
integer scale of map
}
\description{
check name or numeric scale representations, return numeric one
}
\examples{
#commented out because not exported
#check_scale(110)
#check_scale("small")
}

