% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{coastline10}
\alias{coastline10}
\title{world coastlines from Natural Earth at 1:10m (large) resolution}
\format{A \code{SpatialLinesDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/10m/physical/ne_10m_coastline.zip}
}
\usage{
data(coastline10)
}
\description{
world coastlines from Natural Earth at 1:10m (large) resolution
}
\keyword{dataset}

