% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{coastline110}
\alias{coastline110}
\title{world coastlines from Natural Earth at 1:110m (small) resolution}
\format{A \code{SpatialLinesDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/110m/physical/ne_110m_coastline.zip}
}
\usage{
data(coastline110)
}
\description{
world coastlines from Natural Earth at 1:110m (small) resolution
}
\keyword{dataset}

