% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{coastline50}
\alias{coastline50}
\title{world coastlines from Natural Earth at 1:50m (medium) resolution}
\format{A \code{SpatialLinesDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/50m/physical/ne_50m_coastline.zip}
}
\usage{
data(coastline50)
}
\description{
world coastlines from Natural Earth at 1:50m (medium) resolution
}
\keyword{dataset}

