% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{coastlines}
\alias{coastline10}
\alias{coastline110}
\alias{coastline50}
\alias{coastlines}
\alias{coastlines10}
\alias{coastlines110}
\alias{coastlines50}
\title{world coastlines from Natural Earth}
\format{A \code{SpatialLinesDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/110m/physical/ne_110m_coastline.zip}

\url{http//www.naturalearthdata.com/download/50m/physical/ne_50m_coastline.zip}

\url{http//www.naturalearthdata.com/download/10m/physical/ne_10m_coastline.zip}
}
\usage{
coastline110

coastline50

coastline10
}
\description{
coastline lines at scales 1:110m (small), 1:50m (medium), 1:10m (large).
}
\keyword{datasets}

