% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{countries10}
\alias{countries10}
\title{world country polygons from Natural Earth at 1:10m (large) resolution}
\format{A \code{SpatialPolygonsDataFrame} with 255 elements}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_countries.zip}
}
\usage{
data(countries10)
}
\description{
boundaries of world countries and associated attribute data
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with country attributes}
}}
\keyword{dataset}

