% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{countries110}
\alias{countries110}
\title{world country polygons from Natural Earth at 1:110m (small) resolution}
\format{A \code{SpatialPolygonsDataFrame} with 177 obs. of  63 variables}
\source{
\url{http//www.naturalearthdata.com/download/110m/cultural/ne_110m_admin_0_countries.zip}
}
\usage{
data(countries110)
}
\description{
boundaries of world countries and associated attribute data
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with country attributes}
}}
\keyword{dataset}

