% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{countries50}
\alias{countries50}
\title{world country polygons from Natural Earth at 1:50m (medium) resolution}
\format{A \code{SpatialPolygonsDataFrame} with	241 elements}
\source{
\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_countries.zip}
}
\usage{
data(countries50)
}
\description{
boundaries of world countries and associated attribute data
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with country attributes}
}}
\keyword{dataset}

