% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{df_layers_cultural}
\alias{df_layers_cultural}
\title{list of cultural layers available from Natural Earth}
\format{
A \code{DataFrame}

An object of class \code{data.frame} with 43 rows and 4 columns.
}
\usage{
df_layers_cultural
}
\description{
list of cultural layers available from Natural Earth
}
\keyword{datasets}
