% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{df_layers_physical}
\alias{df_layers_physical}
\title{list of physical layers available from Natural Earth}
\format{
A \code{DataFrame}

An object of class \code{data.frame} with 29 rows and 4 columns.
}
\usage{
df_layers_physical
}
\description{
list of physical layers available from Natural Earth
}
\keyword{datasets}
