% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{find_vector_data}
\alias{find_vector_data}
\title{Return a dataframe of available vector layers on Natural Earth}
\usage{
find_vector_data(scale = 110, category = c("cultural", "physical"))
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10} or \code{'small'}, \code{'medium'}, \code{'large'}}

\item{category}{one of natural earth categories : 'cultural', 'physical'}
}
\value{
dataframe with two variables: layers and metadata
}
\description{
Checks the Natural Earth Github repository for current vector layers and
provides the file name required in the type argument of ne_download.
}
\examples{
\dontrun{
find_vector_data(scale = 10, category = "physical")
}
}
