% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{get_data}
\alias{get_data}
\title{Get data from within the package}
\usage{
get_data(scale = 110, type = "countries")
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{'small'}, \code{'medium'}}

\item{type}{country type, one of 'countries', 'map_units', 'sovereignty'}
}
\value{
A \code{SpatialPolygonsDataFrame} object.
}
\description{
returns world country polygons at a specified scale, used by ne_countries()
}

