% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{git_contents}
\alias{git_contents}
\title{Return contents of Natural Earth Github directory}
\usage{
git_contents(path)
}
\arguments{
\item{path}{string, one of: \code{'110m_physical'}, \code{'110m_cultural'},
\code{'50m_physical'}, \code{'50m_cultural'}, \code{'10m_physical'},
\code{'10m_cultural'}}
}
\value{
list. Includes parsed json content, http path, and response
  code.
}
\description{
Uses the Github API to return contents of Natural Earth Github directories.
}
\keyword{internal}
