% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{git_layer_names}
\alias{git_layer_names}
\title{Create list of layer names and metadata links}
\usage{
git_layer_names(x, scale)
}
\arguments{
\item{x}{object returned by git_contents}

\item{scale}{one of \code{110}, \code{50}, \code{10}}
}
\value{
list of lists with layer names and metadata links.
}
\description{
Parses Natural Earth Github folder content for layer names and metadata
links.
}
\keyword{internal}
