% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-rnaturalearthhires.R
\name{install_rnaturalearthhires}
\alias{install_rnaturalearthhires}
\title{Install the naturalearthhires package after checking with the user}
\usage{
install_rnaturalearthhires()
}
\description{
Install the naturalearthhires package after checking with the user
}
