% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{map_units}
\alias{map_units}
\alias{map_units10}
\alias{map_units110}
\alias{map_units50}
\title{world map_unit polygons from Natural Earth}
\format{A \code{SpatialPolygonsDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/110m/cultural/ne_110m_admin_0_map_units.zip}

\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_map_units.zip}

\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_map_units.zip}
}
\usage{
map_units110

map_units50

map_units10
}
\description{
map_unit polygons at scales 1:110m (small), 1:50m (medium), 1:10m (large).
There are more map_units than countries e.g. United Kingdom is split into England, Scotland, Wales and Northern Ireland.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}
\keyword{datasets}

