% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{map_units10}
\alias{map_units10}
\title{world map_unit polygons from Natural Earth at 1:10m (large) resolution}
\format{A \code{SpatialPolygonsDataFrame} with 295 elements}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_map_units.zip}
}
\usage{
data(map_units10)
}
\description{
there are more map_units than countries e.g. French Guiana is separate from France
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}
\keyword{dataset}

