% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{map_units110}
\alias{map_units110}
\title{world map_unit polygons from Natural Earth at 1:110m (small) resolution}
\format{A \code{SpatialPolygonsDataFrame} with 183 elements}
\source{
\url{http//www.naturalearthdata.com/download/110m/cultural/ne_110m_admin_0_map_units.zip}
}
\usage{
data(map_units110)
}
\description{
there are more map_units than countries e.g. French Guiana is separate from France
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}
\keyword{dataset}

