% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_as_sp.R
\name{ne_as_sp}
\alias{ne_as_sp}
\title{coerce return object to sp if option set}
\usage{
ne_as_sp(x, returnclass = c("sp", "sf"))
}
\arguments{
\item{x}{scale of map to return, one of \code{110}, \code{50}, \code{10} or
\code{'small'}, \code{'medium'}, \code{'large'}}

\item{returnclass}{'sp' default or 'sf' for Simple Features}
}
\value{
an sf or sp object
}
\description{
coerce return object to sp if option set
}
