% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data.R
\name{ne_git_layer_names}
\alias{ne_git_layer_names}
\title{Create list of layer names and metadata links}
\usage{
ne_git_layer_names(x, scale, getmeta)
}
\arguments{
\item{x}{object returned by ne_git_contents}

\item{scale}{one of \code{110}, \code{50}, \code{10}}

\item{getmeta}{whether to get url of the metadata for each layer}
}
\value{
list of lists with layer names and metadata links.
}
\description{
Parses Natural Earth Github folder content for layer names and metadata
links.
}
\keyword{internal}
