% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_load.r
\name{ne_load}
\alias{ne_load}
\title{load a Natural Earth vector that has already been downloaded into R}
\usage{
ne_load(scale = 110, type = "countries", category = c("cultural",
  "physical"), destdir = tempdir(), file_name = NULL)
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10}}

\item{type}{type of natural earth file to download one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
OR the portion of any natural earth vector url after the scale and before the . 
e.g. for "ne_50m_urban_areas.zip" this would be "urban_areas"}

\item{category}{one of natural earth categories : 'cultural', 'physical'}

\item{destdir}{folder to load files from, default=tempdir()}

\item{file_name}{OPTIONAL name of file (excluding path) instead of natural earth attributes}
}
\value{
A \code{Spatial} object depending on the vector source (points, lines or polygons).
}
\description{
returns loaded vector as a spatial object.
}
\examples{
#commented out to stop download in check
#download followed by load from tempdir() works in same R session
#spdf_world <- ne_download( scale = 110, type = 'countries' )
#spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
#download followed by load from specified directory works between R sessions
#spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
#spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
}

