% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_file_name.R
\name{normalize_type}
\alias{normalize_type}
\title{Normalize the type argument for Natural Earth datasets}
\usage{
normalize_type(type)
}
\arguments{
\item{type}{type of natural earth file to download one of 'countries',
'map_units', 'map_subunits', 'sovereignty', 'states' OR the portion of any
natural earth vector url after the scale and before the . e.g. for
'ne_50m_urban_areas.zip' this would be 'urban_areas' OR the raster
filename e.g. for 'MSR_50M.zip' this would be 'MSR_50M'}
}
\value{
A string representing the normalized dataset type.
}
\description{
This function standardizes the `type` argument by mapping common names to
their respective Natural Earth dataset names.
}
