% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_spatial_vector}
\alias{read_spatial_vector}
\title{Read Spatial Vector}
\usage{
read_spatial_vector(x, layer, returnclass = c("sf", "sv"))
}
\arguments{
\item{x}{A character string specifying the path to the spatial vector file.}

\item{layer}{A character string specifying the later to read in case there
are more than one in the zip file.}

\item{returnclass}{A character string specifying the class of object to
return. Options are "sf" for sf object and "sv" for sv object.}
}
\value{
Either an `sf` object or a `sv` object.
}
\description{
This function reads a spatial vector file and returns either an `sf` object
or a `sv` object.
}
