% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaturalearth-package.R
\docType{package}
\name{rnaturalearth-package}
\alias{rnaturalearth}
\alias{rnaturalearth-package}
\title{rnaturalearth: World Map Data from Natural Earth}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Facilitates mapping by making natural earth map data from \url{https://www.naturalearthdata.com/} more easily available to R users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/rnaturalearth/}
  \item \url{https://github.com/ropensci/rnaturalearth}
  \item Report bugs at \url{https://github.com/ropensci/rnaturalearth/issues}
}

}
\author{
\strong{Maintainer}: Philippe Massicotte \email{pmassicotte@hotmail.com} (\href{https://orcid.org/0000-0002-5919-4116}{ORCID})

Authors:
\itemize{
  \item Andy South \email{southandy@gmail.com}
}

Other contributors:
\itemize{
  \item Koen Hufkens \email{koen.hufkens@gmail.com} (\href{https://orcid.org/0000-0002-5070-8109}{ORCID}) [contributor]
}

}
\keyword{internal}
