% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_gdal_url}
\alias{sanitize_gdal_url}
\title{Extracts the http URL from a VSIZIP URL}
\usage{
sanitize_gdal_url(url)
}
\arguments{
\item{url}{A character string representing the VSIZIP URL}
}
\value{
A character string representing the http URL extracted from the
VSIZIP URL
}
\description{
This function takes a VSIZIP URL and extracts the http URL from
it
}
