% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sovereignty110}
\alias{sovereignty110}
\title{world sovereignty polygons from Natural Earth at 1:110m (small) resolution}
\format{A \code{SpatialPolygonsDataFrame} with 171 elements}
\source{
\url{http//www.naturalearthdata.com/download/110m/cultural/ne_110m_admin_0_sovereignty.zip}
}
\usage{
data(sovereignty110)
}
\description{
there are fewer sovereign states than countries e.g. the Falkland Islands are included with the United Kingdom
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}
\keyword{dataset}

