% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sovereignty50}
\alias{sovereignty50}
\title{world sovereignty polygons from Natural Earth at 1:50m (medium) resolution}
\format{A \code{SpatialPolygonsDataFrame} with 200 elements}
\source{
\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_sovereignty.zip}
}
\usage{
data(sovereignty50)
}
\description{
there are fewer sovereign states than countries e.g. the Falkland Islands are included with the United Kingdom
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}
\keyword{dataset}

