% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{states}
\alias{states}
\alias{states10}
\alias{states50}
\title{state (admin level 1) polygons from Natural Earth}
\format{A \code{SpatialPolygonsDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_1_states.zip}

\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_1_states.zip}
}
\usage{
states10

states50
}
\description{
For all world countries at 1:10m (high) resolution, for Australia, Brazil, Canada and USA, from Natural Earth at 1:50m (medium) resolution
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with attributes}
}}
\keyword{datasets}

