% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{states10}
\alias{states10}
\title{world state (admin level 1) polygons from Natural Earth at 1:10m (high) resolution}
\format{A \code{SpatialPolygonsDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_1_states.zip}
}
\usage{
data(states10)
}
\description{
world state (admin level 1) polygons from Natural Earth at 1:10m (high) resolution
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with state attributes}
}}
\keyword{dataset}

