% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{states50}
\alias{states50}
\title{state (admin level 1) polygons for Australia, Brazil, Canada and USA, from Natural Earth at 1:50m (medium) resolution}
\format{A \code{SpatialPolygonsDataFrame}}
\source{
\url{http//www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_1_states.zip}
}
\usage{
data(states50)
}
\description{
state (admin level 1) polygons for Australia, Brazil, Canada and USA, from Natural Earth at 1:50m (medium) resolution
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with state attributes}
}}
\keyword{dataset}

