% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/world_countries.r
\name{world_countries}
\alias{world_countries}
\title{Get world country polygons}
\usage{
world_countries(scale = 110, type = "countries", continent = NULL,
  country = NULL, geounit = NULL, sovereignty = NULL)
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{'small'}, \code{'medium'}}

\item{type}{country type, one of 'countries', 'map_units', 'sovereignty'}

\item{continent}{a character vector of continent names to get countries from.}

\item{country}{a character vector of country names.}

\item{geounit}{a character vector of geounit names.}

\item{sovereignty}{a character vector of sovereignty names.}
}
\value{
A \code{SpatialPolygonsDataFrame} object.
}
\description{
returns world country polygons at a specified scale
}
\examples{
spdf_world <- world_countries()
spdf_africa <- world_countries(continent='africa')
spdf_france <- world_countries(country='france')

if (require(sp)) {
  plot(spdf_world)
  plot(spdf_africa)
  plot(spdf_france)
}
}

