% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_states.r
\name{world_states}
\alias{world_states}
\title{Get world state (admin level 1) polygons}
\usage{
world_states(country = NULL, geounit = NULL, iso_a2 = NULL, spdf = NULL)
}
\arguments{
\item{country}{a character vector of country names.}

\item{geounit}{a character vector of geounit names.}

\item{iso_a2}{a character vector of iso_a2 country codes}

\item{spdf}{an optional alternative states map}
}
\value{
A \code{SpatialPolygonsDataFrame} object.
}
\description{
returns world state polygons (administrative level 1) for specified countries
}
\examples{

#comparing using country and geounit to filter
spdf_france_country <- world_states(country='france')
spdf_france_geounit <- world_states(geounit='france')
 
if (require(sp)) {
  plot(spdf_france_country)
  plot(spdf_france_geounit) 
  
  plot(world_states(country ='united kingdom'))  
  plot(world_states(geounit='england'))     
}
}

