% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-rnaturalearthhires.r
\name{check_rnaturalearthhires}
\alias{check_rnaturalearthhires}
\title{Check whether to install rnaturalearthhires and install if necessary}
\usage{
check_rnaturalearthhires()
}
\description{
If the rnaturalearthhires package is not installed, install it from GitHub using
devtools. If it is not up to date, reinstall it.
}

