% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{countries}
\alias{countries}
\alias{countries110}
\title{world country polygons from Natural Earth}
\format{
A \code{SpatialPolygonsDataFrame}

An object of class \code{SpatialPolygonsDataFrame} with 177 rows and 94 columns.
}
\source{
\url{https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_0_countries.zip}
}
\usage{
countries110
}
\description{
at 1:110m scale (small). Other data and resolutions are in the packages
rnaturalearthdata and rnaturalearthhires.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data frame with country attributes}
}}

\keyword{datasets}
