% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_countries.r
\name{ne_countries}
\alias{ne_countries}
\title{Get natural earth world country polygons}
\usage{
ne_countries(scale = 110, type = "countries", continent = NULL,
  country = NULL, geounit = NULL, sovereignty = NULL)
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{'small'}, \code{'medium'}}

\item{type}{country type, one of 'countries', 'map_units', 'sovereignty'}

\item{continent}{a character vector of continent names to get countries from.}

\item{country}{a character vector of country names.}

\item{geounit}{a character vector of geounit names.}

\item{sovereignty}{a character vector of sovereignty names.}
}
\value{
A \code{SpatialPolygonsDataFrame} object.
}
\description{
returns world country polygons at a specified scale
}
\examples{
spdf_world <- ne_countries()
spdf_africa <- ne_countries(continent='africa')
spdf_france <- ne_countries(country='france')

if (require(sp)) {
  plot(spdf_world)
  plot(spdf_africa)
  plot(spdf_france)   
}
}

