% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_countries.r
\name{ne_countries}
\alias{ne_admin0}
\alias{ne_countries}
\title{Get natural earth world country polygons}
\usage{
ne_countries(scale = 110, type = "countries", continent = NULL,
  country = NULL, geounit = NULL, sovereignty = NULL)
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10} or \code{'small'}, \code{'medium'}, \code{'large'}}

\item{type}{country type, one of 'countries', 'map_units', 'sovereignty', 'tiny_countries'}

\item{continent}{a character vector of continent names to get countries from.}

\item{country}{a character vector of country names.}

\item{geounit}{a character vector of geounit names.}

\item{sovereignty}{a character vector of sovereignty names.}
}
\value{
\code{SpatialPolygonsDataFrame} or \code{SpatialPointsDataFrame} object.
}
\description{
returns world country polygons at a specified scale, or points of tiny_countries
}
\examples{
spdf_world <- ne_countries()
spdf_africa <- ne_countries(continent = 'africa')
spdf_france <- ne_countries(country = 'france')
spdf_tiny_countries <- ne_countries(type = 'tiny_countries', scale = 50)

if (require(sp)) {
  plot(spdf_world)
  plot(spdf_africa)
  plot(spdf_france)
  plot(spdf_tiny_countries)   
}
}

