% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ne_download.r
\name{ne_download}
\alias{ne_download}
\title{download a vector from Natural Earth and (optionally) read into R}
\usage{
ne_download(scale = 110, type = "countries", category = c("cultural",
  "physical", "raster"), destdir = tempdir())
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10}}

\item{type}{type of natural earth file to download one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
OR the portion of any natural earth vector url after the scale and before the .
e.g. for "ne_50m_urban_areas.zip" this would be "urban_areas"}

\item{category}{one of natural earth categories : 'cultural', 'physical', 'raster'}

\item{destdir}{where to save files, defaults to \code{tempdir()}}
}
\value{
A \code{Spatial} object depending on the vector source (points, lines or polygons).
}
\description{
returns downloaded vector as a spatial object.
}
\examples{
spdf_world <- ne_download( scale = 110, type = 'countries' )

if (require(sp)) {
  plot(spdf_world)
  plot(ne_download(type='populated_places'))
}
#to save downloaded files locally use destdir
#spdf_world <- ne_download( scale = 110, type = 'countries', destdir=getwd() )
#then ne_load() with the same args can be used to access the saved file
#spdf_world_from_saved <- ne_load( scale = 110, type = 'countries', destdir=getwd() )
}

