% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_download.r
\name{ne_download}
\alias{ne_download}
\title{download data from Natural Earth and (optionally) read into R}
\usage{
ne_download(scale = 110, type = "countries", category = c("cultural",
  "physical", "raster"), destdir = tempdir(), load = TRUE)
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10}}

\item{type}{type of natural earth file to download one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
OR the portion of any natural earth vector url after the scale and before the . 
e.g. for "ne_50m_urban_areas.zip" this would be "urban_areas"}

\item{category}{one of natural earth categories : 'cultural', 'physical', 'raster'}

\item{destdir}{where to save files, defaults to \code{tempdir()}}

\item{load}{TRUE/FALSE whether to load file into R and return}
}
\value{
A \code{Spatial} object depending on the data (points, lines, polygons or raster), 
   unless load=FALSE in which case it returns the name of the downloaded shapefile (without extension).
}
\description{
returns downloaded data as a spatial object or the filename if \code{load=FALSE}.  
if \code{destdir} is specified the data can be reloaded in a later R session using \code{\link{ne_load}}
with the same arguments.
}
\examples{
spdf_world <- ne_download( scale = 110, type = 'countries' )

if (require(sp)) {
  plot(spdf_world)
  plot(ne_download(type='populated_places'))
}

#reloading from the saved file in the same session with same arguments
spdf_world2 <-    ne_load( scale = 110, type = 'countries' )

#download followed by load from specified directory will work across sessions
#spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
#spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )

#' #for raster
#download & load
#rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
#load after having downloaded
#rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
#plot
#library(raster)
#plot(rst)
 
}
\seealso{
\code{\link{ne_load}}
}

