% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_load.r
\name{ne_load}
\alias{ne_load}
\title{load a Natural Earth vector that has already been downloaded to R using \code{\link{ne_download}}}
\usage{
ne_load(scale = 110, type = "countries", category = c("cultural",
  "physical", "raster"), destdir = tempdir(), file_name = NULL)
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10} or \code{'small'}, \code{'medium'}, \code{'large'}}

\item{type}{type of natural earth file one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
OR the portion of any natural earth vector url after the scale and before the . 
e.g. for "ne_50m_urban_areas.zip" this would be "urban_areas"}

\item{category}{one of natural earth categories : 'cultural', 'physical', 'raster'}

\item{destdir}{folder to load files from, default=tempdir()}

\item{file_name}{OPTIONAL name of file (excluding path) instead of natural earth attributes}
}
\value{
A \code{Spatial} object depending on the data (points, lines, polygons or raster).
}
\description{
returns loaded data as a spatial object.
}
\examples{
# commented out to stop download in check
# download followed by load from tempdir() works in same R session
# spdf_world <- ne_download( scale = 110, type = 'countries' )
# spdf_world2 <-    ne_load( scale = 110, type = 'countries' )
# download followed by load from specified directory works between R sessions
# spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
# spdf_world2 <-    ne_load( scale = 110, type = 'countries', destdir = getwd() )
# for raster
# download & load
# rst <- ne_download(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
# load after having downloaded
# rst <- ne_load(scale = 50, type = "OB_50M", category = "raster", destdir = getwd())
# plot
# library(raster)
# plot(rst)
}
\seealso{
\code{\link{ne_download}}
}

