% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ne_load.r
\name{ne_load}
\alias{ne_load}
\title{load a Natural Earth vector that has already been downloaded into R}
\usage{
ne_load(scale = 110, type = "countries", category = c("cultural",
  "physical", "raster"), destdir = NULL, file_name = NULL)
}
\arguments{
\item{scale}{scale of map to return, one of \code{110}, \code{50}, \code{10}}

\item{type}{type of natural earth file to download one of 'countries', 'map_units', 'map_subunits', 'sovereignty', 'states'
OR the portion of any natural earth vector url after the scale and before the .
e.g. for "ne_50m_urban_areas.zip" this would be "urban_areas"}

\item{category}{one of natural earth categories : 'cultural', 'physical', 'raster'}

\item{destdir}{folder to load files from}

\item{file_name}{OPTIONAL name of file (excluding path) instead of natural earth attributes}
}
\value{
A \code{Spatial} object depending on the vector source (points, lines or polygons).
}
\description{
returns loaded vector as a spatial object.
}
\examples{
#commented out to stop download in check
#download to wd
#spdf_world <- ne_download( scale = 110, type = 'countries', destdir = getwd() )
#to load from wd later
#spdf_world <- ne_load( scale = 110, type = 'countries', destdir = getwd() )
}

