% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaturalearth-package.r
\docType{package}
\name{rnaturalearth}
\alias{rnaturalearth}
\alias{rnaturalearth-package}
\title{rnaturalearth : world vector maps from Natural Earth}
\description{
Facilitates world mapping by making \href{http://www.naturalearthdata.com/}{Natural Earth} map data more easily available to R users. Focuses on vector data.
}
\seealso{
ne_countries
}

